+++ tests/test_objects.py
@@ -187,6 +187,14 @@ def test_invert_by(case, expected):
 
 
 @parametrize('case,expected', [
+    (({'a': 1, 'b': 2}, 'get', 'a'), 1),
+    (({'a': {'b': {'c': [1, 2, 3, 4]}}}, 'a.b.c.copy'), [1, 2, 3, 4]),
+])
+def test_invoke(case, expected):
+    assert _.invoke(*case) == expected
+
+
+@parametrize('case,expected', [
     # NOTE: The expected is a list of values but find_key returns only a single
     # value. However, since dicts do not have an order, it's unknown what the
     # "first" returned value will be.